Move addNodeOnGraph to litegraphService
